refactor(studio): move ai + runtime infra out of modules/ (Tauri-shell step 2a)
Continue reshaping src/. Relocate the two genuine infrastructure subsystems out
of the generic `modules/` wrapper to their own top-level homes (cohesive
subsystems, like inference.rs/gpu.rs — not business-module folders):
- `src/modules/ai/` -> `src/ai/` (AiService + llm/sam/runtime_setup/plugin/
  registry/engines; the ONNX/HTTP/file infra pending crate extraction).
- `src/modules/runtime/` -> `src/runtime/` (the runtime glue).

Rewrote the paths (`crate::modules::ai::`->`crate::ai::`,
`crate::modules::runtime::`->`crate::runtime::`), moved the `pub mod` decls from
`modules/mod.rs` to lib.rs. `modules/` now holds only the 6 thin migrated-module
facades/shims (projects/labels/images/cloud/video/analysis), which dissolve into
src/composition next. Green: default `cargo check`.
