+++ apps/studio/src-tauri/src/ai/engines/mod.rs
@@ -1,7 +1,7 @@
 //! Concrete inference engines for capability-aware model plugins.
 //!
-//! Each engine implements [`crate::modules::ai::plugin::ModelPlugin`] and is wired
-//! into [`crate::modules::ai::plugin::plugin_for`]. They are gated behind the
+//! Each engine implements [`crate::ai::plugin::ModelPlugin`] and is wired
+//! into [`crate::ai::plugin::plugin_for`]. They are gated behind the
 //! `local-inference` feature (the ONNX Runtime stack); without it the registry
 //! entries resolve to the scaffold `NotImplementedPlugin`.
