+++ apps/studio/src-tauri/src/ai/mod.rs
