+++ apps/studio/src-tauri/src/ai/registry.rs
@@ -2,7 +2,7 @@
 //!
 //! The local model registry (the catalog of known models + their capabilities)
 //! now lives in `vailabel_models::domain::registry`. This shim keeps the paths
-//! `crate::modules::ai::registry::{RegistryModel, REGISTRY, registry_json, find}`
+//! `crate::ai::registry::{RegistryModel, REGISTRY, registry_json, find}`
 //! valid so `commands.rs` (the `ai_model_registry` command) and `plugin.rs` (the
 //! dispatcher's `super::registry::find`) compile unchanged.
