+++ apps/studio/src-tauri/src/ai/runtime_setup.rs
