+++ apps/studio/src-tauri/src/modules/mod.rs
@@ -1,8 +1,6 @@
-pub mod ai;
 pub mod analysis;
 pub mod cloud;
 pub mod images;
 pub mod labels;
 pub mod projects;
-pub mod runtime;
 pub mod video;
